Moving credential provider tests about
