--- Tests/AWSSDKSwiftCoreTests/Credential/StaticCredential+CredentialProviderTests.swift
@@ -1,31 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the AWSSDKSwift open source project
-//
-// Copyright (c) 2017-2020 the AWSSDKSwift project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of AWSSDKSwift project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-@testable import AWSSDKSwiftCore
-import XCTest
-import NIO
-
-class StaticCredential_CredentialProviderTests: XCTestCase {
-    func testCredentialProvider() {
-        let cred = StaticCredential(accessKeyId: "abc", secretAccessKey: "123", sessionToken: "xyz")
-        
-        let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-        defer { XCTAssertNoThrow(try group.syncShutdownGracefully()) }
-        let loop = group.next()
-        var returned: Credential?
-        XCTAssertNoThrow(returned = try cred.getCredential(on: loop).wait())
-        
-        XCTAssertEqual(returned as? StaticCredential, cred)
-    }
-}
